happy path for adding host and help request by a trusted user
