+++ app/Http/Controllers/ShareController.php
@@ -2,15 +2,26 @@
 
 namespace App\Http\Controllers;
 
+use A17\Twill\Repositories\SettingRepository;
 use App\Accommodation;
 use App\AccommodationType;
 use App\Country;
 use App\County;
 use App\FacilityType;
+use App\HelpRequest;
 use App\Http\Requests\AccommodationRequest;
+use App\Http\Requests\ServiceRequest;
+use App\Language;
+use App\Mail\HelpRequestMail;
 use App\Services\AccommodationService;
+use App\Services\HelpRequestService;
+use App\Services\UserService;
+use App\UaRegion;
 use App\User;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\App;
+use Illuminate\Support\Facades\Mail;
 
 class ShareController extends Controller
 {
@@ -45,8 +56,6 @@ class ShareController extends Controller
             ->with('countries', Country::all())
             ->with('counties', County::all())
             ->with('hostType', old('host_type_copy'))
-            ->with('countries', Country::all())
-            ->with('counties', County::all())
             ->with('description', '');;
 
     }
@@ -67,5 +76,58 @@ class ShareController extends Controller
         return redirect()->back();
 
 
+    }
+
+    public function helpRequestList(Request $request)
+    {
+        $user = auth()->user();
+        return view('admin.help-list',)
+            ->with('approvalStatus', $request->get('status'))
+            ->with('user', $user);
+    }
+
+    public function helpRequestCreate(Request $request, SettingRepository $settingRepository)
+    {
+        $user = null;
+        if (session()->get('createdRefugeeUserId'))
+        {
+            $user = User::find(session()->get('createdRefugeeUserId'));
+        }
+        $languages = Language::orderBy('position', 'asc')->orderBy('name', 'asc')->select('id', 'endonym')->get();
+
+        $lang = App::getLocale() == 'ro' ? 'en' : App::getLocale();
+        $counties = UaRegion::all(['id', 'region', 'region_' . $lang . ' as region'])->sortBy('region_' . $lang);
+
+
+        return view('share.help-request-add')
+            ->with('description', $settingRepository->byKey('request_services_description') ?? '')
+            ->with('info', $settingRepository->byKey('request_services_info') ?? '')
+            ->with('user', $user)
+            ->with('counties', $counties)
+            ->with('hostType', 'person')
+            ->with('languages', $languages);
+
+    }
+
+    public function helpRequestStore(Request $request)
+    {
+        $helpRequestService = new HelpRequestService();
+        if (session()->get('createdRefugeeUserId'))
+        {
+            $sessionUserId = session()->pull('createdRefugeeUserId');
+            $user = User::find($sessionUserId);
+            $helpRequestService->create($request,$user,auth()->user()->id);
+            session()->flash('status',__('Help request created successfully'));
+            return redirect()->route('share.help.request.list');
+        }
+        session()->flash('status',__('Refugee user not found'));
+        return redirect()->back();
+    }
+
+    public function createHelpRequestUser(ServiceRequest $request): RedirectResponse
+    {
+        $user = (new UserService())->createRefugeeUser($request);
+        session()->put('createdRefugeeUserId',$user->id);
+        return redirect()->back();
     }
 }
