+++ resources/views/layouts/partials/sidebars/trusted.blade.php
@@ -7,6 +7,6 @@
 <a href="{{ route('share.help.request.list') }}" class="list-group-item list-group-item-action ">
     <i class="fa fa-bed mr-3"></i>{{__('Help request')}}
 </a>
-<a href="{{ route('share.accommodation.create', ['status' => 1]) }}" class="list-group-item list-group-item-action sub-list ">
+<a href="{{ route('share.help.request.create', ['status' => 1]) }}" class="list-group-item list-group-item-action sub-list ">
     <i class="fa fa-minus mx-3"></i>{{__('Add help request')}}
 </a>
