+++ resources/views/share/partials/add-user-help-request-modal.blade.php
@@ -1,0 +1,28 @@
+<!-- Button trigger modal -->
+<button type="button" class="btn btn-primary" data-toggle="modal" data-target="#exampleModal">
+    {{__('Create user')}}
+</button>
+@if(!empty($user))<h2>User: {{$user->name}} </h2>@endif
+
+<!-- Modal -->
+<div class="modal fade" id="exampleModal" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel" aria-hidden="true">
+    <div class="modal-dialog" role="document">
+        <div class="modal-content">
+            <div class="modal-header">
+                <h5 class="modal-title" id="exampleModalLabel">{{__('Add user')}}</h5>
+                <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                    <span aria-hidden="true">&times;</span>
+                </button>
+            </div>
+            <div class="modal-body">
+                @include('partials.forms.request-services-add-user', [
+                    'formRoute' => route('share.help.request.create.refugee'),
+                    'showFormHeader'    => false
+                ])
+            </div>
+            <div class="modal-footer">
+                <button type="button" class="btn btn-secondary" data-dismiss="modal">{{__('Close')}}</button>
+            </div>
+        </div>
+    </div>
+</div>
